% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_funs.R
\name{log_like_cpp}
\alias{log_like_cpp}
\title{log_like_cpp - a function to compute the log likelihood for challenge 8}
\usage{
log_like_cpp(params, X, y, tpts)
}
\value{
A scalar
}
\description{
log_like_cpp - a function to compute the log likelihood for challenge 8
}
\details{
More summary needed here.
}
