% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_funs.R
\name{log_post_cpp}
\alias{log_post_cpp}
\title{log_post_cpp - a function to compute the log posterior for challenge 8}
\usage{
log_post_cpp(params, X, y, tpts)
}
\value{
A scalar
}
\description{
log_post_cpp - a function to compute the log posterior for challenge 8
}
\details{
More summary needed here.
}
