% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_funs.R
\name{MH}
\alias{MH}
\title{MH - a function to run MH for challenge 8}
\usage{
MH(MH_draws, proposal_cov, X, y, tpts)
}
\value{
A scalar for number of accepted draws and also modifies the matrix supplied to the first argument in place
}
\description{
MH - a function to run MH for challenge 8
}
\details{
More summary needed here.
}
